% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msigdb_browse.R
\name{msigdb_browse}
\alias{msigdb_browse}
\title{msigdb_browser}
\usage{
msigdb_browse(species, name)
}
\arguments{
\item{species}{A specie, Hu or Mm.}

\item{name}{The gene sets' name to load.}
}
\value{
The gene set's that you selected.
}
\description{
You can load MsigDB gene sets list on R from vector.
}
