% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msigdb_search.R
\name{msigdb_search}
\alias{msigdb_search}
\title{msigdb_search}
\usage{
msigdb_search(species = NULL)
}
\arguments{
\item{species}{A specie, Hu or Mm.}
}
\value{
The gene set's name list that you selected.
}
\description{
You can search MsigDB gene sets on R from new shiny tab.
}
